backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
exceptiongroup==1.3.1
fastapi==0.135.1
h11==0.16.0
httpx==0.28.1
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
python-multipart==0.0.22
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
pypdf==6.1.3
python-docx==1.2.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
redis==5.2.0
qdrant-client==1.12.1
openai==1.58.1
celery==5.4.0
alembic==1.14.1
